Rename SecurityAnalysis2
